pyproject.toml
[project]
name = "heliannuuthus-ai"
version = "0.1.0"
description = "一个 A项目"
authors = [{ name = "heliannutuhus", email = "heliannuthus@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
packages = [{ include = "app", from = "." }]

[tool.poetry.dependencies]
openai = "^1.65.3"
pyyaml = "^6.0.2"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"

[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[tool.poetry.scripts]
start = "app.main:start"

[tool.yapf]
column_limit = 120

[tool.yapfignore]
ignore_patterns = [
  ".venv",
  "logs",
  "**/__pycache__/**",
  "*.pyc"
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
